MdePkg/BaseLib: Add a new VMGEXIT instruction invocation for SVSM
The SVSM specification relies on a specific register calling convention
to hold the parameters that are associated with the SVSM request.

Create a new version of the VMGEXIT instruction that will adhere to this
calling convention and load the SVSM function arguments into the proper
register before invoking the VMGEXIT instruction.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>